server/requirements.txt
amqp==2.5.2
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
billiard==3.6.1.0
black==19.3b0
celery==4.4.0
Click==7.0
coverage==5.0.2
decorator==4.4.1
Django==3.0.2
django-sendgrid-v5==0.8.1
django-cors-headers==3.2.1
graphene-file-upload==1.2.2
django-graphql-social-auth==0.1.4
django-extensions==2.2.5
django-graphql-social-auth[jwt]
social-auth-app-django
django-graphql-jwt==0.3.0
django-jwt-auth==0.0.2
factory-boy==2.12.0
Faker==3.0.0
fastdiff==0.2.0
flower==0.9.3
gevent==1.4.0
graphene==2.1.8
graphene-django==2.8.0
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.15
gunicorn==20.0.4
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
kombu==4.6.7
more-itertools==8.0.2
packaging==20.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pip-review==1.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.7.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
Rx==1.6.1
serious-django-graphene==0.5.1
singledispatch==3.4.0.3
six==1.13.0
snapshottest==0.5.1
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.0
tornado==5.1.1
traitlets==4.3.3
vine==1.3.0
wasmer==0.3.0
wcwidth==0.1.8
requests-oauthlib==1.3.0
channels==2.4.0
graphql-ws-django==0.3.4
requests==2.23.0
fastecdsa==2.1.3
ecdsa==0.15
